Separetes tempplates from components
