+++ src/components/task-edit.js
@@ -1,134 +1,6 @@
-import {DAYS, MONTH_NAMES, COLORS} from '../consts.js';
-import {formatTime, createElement} from '../utils.js';
+import {createElement} from '../utils.js';
+import {createTaskEditTemplate} from '../templates/task-edit.js';
 
-const createColorsMarkup = (colors, currentColor) => {
-  return colors.map((color, index) => {
-    return (
-      `<input
-      type="radio"
-      id="color-${color}-${index}"
-      class="card__color-input card__color-input--${color} visually-hidden"
-      name="color"
-      value="${color}"
-      ${currentColor === color ? `checked` : ``}
-    />
-    <label
-      for="color-${color}-${index}"
-      class="card__color card__color--${color}"
-      >${color}</label
-    >`
-    );
-  })
-    .join(`\n`);
-};
-
-const createReapitingDaysMarkup = (days, repeatingDays) => {
-  return days.map((day, index) => {
-    const isChecked = repeatingDays[day];
-    return (
-      `<input
-        class="visually-hidden card__repeat-day-input"
-        type="checkbox"
-        id="repeat-mo-4"
-        name="repeat"
-        value="mo"
-        ${isChecked ? `checked` : ``}
-      />
-      <label class="card__repeat-day" for="repeat-${day}-${index}"
-        >${day}</label
-      >`
-    );
-  });
-
-};
-
-const createTaskEditTemplate = (task) => {
-  const {description, color, dueDate, repeatingDays} = task;
-
-  const isExpired = dueDate instanceof Date && dueDate < Date.now();
-  const isDateShowing = !!dueDate;
-
-  const date = isDateShowing ? `${dueDate.getDate()} ${MONTH_NAMES[dueDate.getMonth()]}` : ``;
-  const time = isDateShowing ? formatTime(dueDate) : ``;
-
-  const isRepeatingTask = Object.values(repeatingDays).some(Boolean);
-  const repeatClass = isRepeatingTask ? `card--repeat` : ``;
-  const deadlineClass = isExpired ? `card--deadline` : ``;
-
-  const colorsMarkup = createColorsMarkup(COLORS, color);
-  const repeatingDaysMarkup = createReapitingDaysMarkup(DAYS, repeatingDays);
-
-  return (
-    `<article class="card card--edit card--${color} ${repeatClass} ${deadlineClass}">
-            <form class="card__form" method="get">
-              <div class="card__inner">
-                <div class="card__color-bar">
-                  <svg class="card__color-bar-wave" width="100%" height="10">
-                    <use xlink:href="#wave"></use>
-                  </svg>
-                </div>
-
-                <div class="card__textarea-wrap">
-                  <label>
-                    <textarea
-                      class="card__text"
-                      placeholder="Start typing your text here..."
-                      name="text"
-                    >${description}</textarea>
-                  </label>
-                </div>
-
-                <div class="card__settings">
-                  <div class="card__details">
-                    <div class="card__dates">
-                      <button class="card__date-deadline-toggle" type="button">
-                        date: <span class="card__date-status">${isDateShowing ? `yes` : `no`}</span>
-                      </button>
-                      ${isDateShowing ?
-      `<fieldset class="card__date-deadline">
-                            <label class="card__input-deadline-wrap">
-                              <input
-                                class="card__date"
-                                type="text"
-                                placeholder=""
-                                name="date"
-                                value="${date} ${time}"
-                              />
-                            </label>
-                      </fieldset>`
-      : ``}
-
-                      <button class="card__repeat-toggle" type="button">
-                        repeat:<span class="card__repeat-status">${isRepeatingTask ? `yes` : `no`}</span>
-                      </button>
-
-                    ${isRepeatingTask ? `<fieldset class="card__repeat-days">
-                        <div class="card__repeat-days-inner">
-                          ${repeatingDaysMarkup}
-                        </div>
-                      </fieldset>`
-      : ``}
-
-                    </div>
-                  </div>
-
-                  <div class="card__colors-inner">
-                    <h3 class="card__colors-title">Color</h3>
-                    <div class="card__colors-wrap">
-                      ${colorsMarkup}
-                    </div>
-                  </div>
-                </div>
-
-                <div class="card__status-btns">
-                  <button class="card__save" type="submit">save</button>
-                  <button class="card__delete" type="button">delete</button>
-                </div>
-              </div>
-            </form>
-          </article>`
-  );
-};
 
 export default class TaskEdit {
   constructor(task) {
