+++ src/templates/no-task.js
@@ -1,0 +1,10 @@
+const creteNoTaskTemplate = () => {
+  return (
+    `<p class="board__no-tasks">
+      Click «ADD NEW TASK» in menu to create your first task
+    </p>`
+  );
+};
+
+
+export {creteNoTaskTemplate};
