+++ src/templates/sort.js
@@ -1,0 +1,9 @@
+const createSortTemplate = () => {
+  return (`<div class="board__filter-list">
+        <a href="#" class="board__filter">SORT BY DEFAULT</a>
+        <a href="#" class="board__filter">SORT BY DATE up</a>
+        <a href="#" class="board__filter">SORT BY DATE down</a>
+    </div>`);
+};
+
+export {createSortTemplate};
